Ensure that replica is ahead of replication slot
There is a replication lag between a primary and replica database.
We get information about what data the replication slot can provide
from the primary but perform the initial data snapshot on the replica.
Therefore, we must make sure that the replica has caught up to at
least the LSN which we are going to begin streaming from the
replication slot.